Introduce petab.v1 package (#282)
As discussed in #271, we need to accommodate the changes related to the upcoming PEtab v2. The goal is to move PEtab 1.0 functionality to a `petab.v1` subpackage and PEtab 2.0 functionality to a `petab.v2` subpackage.

This PR moves (almost) all code from the `petab` package to a `petab.v1` sub-package. Keeps all non-private objects importable from their previous location, but issues `DeprecationWarnings`.

**From the next release on, all consumers should change all `from petab[.$x] import $y` to `from petab.v1[.$x] import $y`.**

Fixes a couple of sphinx-issues that occurred on the way. Some obscure sphinx-failures remain, but they aren't critical.

Also adds some missing `__all__`s.