Function for replacing `objectivePrior*` by observables/measurements
Reformulate the given PEtab problem such that the objective priors are converted
to measurements. This is done by adding a new observable for each prior
and adding a corresponding measurement to the measurement table.
The resulting optimization problem will be equivalent to the original problem.
This is meant to be used for tools that do not support priors.

Closes #307