petab.calculate: compare all common columns
For computing residuals, ... from measurement + simulation tables,
we need to match the corresponding rows. Previously, this was done
using a subset of PEtab measurement table columns and checking
whether all values in these columns match.

This changes it to using the full set of overlapping columns,
not only the known measurement columns.
With that, the same functions can be used for PEtab v2
measurement/simulation tables.