Add LogUniform distribution (#465)
Add an explicit LogUniform distribution class.

The interpretation of the distribution parameter is different from the existing `Uniform(a, b, log=True)`.

In PEtab v2, X ~ LogUniform(a, b) <=> ln(X) ~ Uniform(ln(a), ln(b)).

However, in PEtab v1, a `parameterScaleUniform` prior for a parameterScale=log parameter is interpreted as ln(X) ~ Uniform(a, b).